Move dynamic user page to its own folder
